split coffee.system and switch to absolute imports
