+++ coffee/expander.py
@@ -31,11 +31,13 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
 # OF THE POSSIBILITY OF SUCH DAMAGE.
 
+from __future__ import absolute_import
+
 import itertools
 
-from base import *
-from utils import *
-from exceptions import UnexpectedNode
+from .base import *
+from .utils import *
+from .exceptions import UnexpectedNode
 
 
 class Expander(object):
