+++ coffee/factorizer.py
@@ -31,10 +31,13 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
 # OF THE POSSIBILITY OF SUCH DAMAGE.
 
+from __future__ import absolute_import
+
 import operator
 
-from base import *
-from utils import *
+from .base import *
+from .utils import *
+from functools import reduce
 
 
 class Term(object):
